optimize configuration file

+++ ciweimao/common.go
@@ -30,16 +30,16 @@ type bookTocStruct struct {
 	Chapters []structure.ChapterList
 }
 
-func splitArray(arr []structure.ChapterList, num int64) [][]structure.ChapterList {
+func splitArray(arr []structure.ChapterList, num int) [][]structure.ChapterList {
 	var segmens = make([][]structure.ChapterList, 0)
-	max := int64(len(arr))
+	max := len(arr)
 	if max < num {
 		segmens = append(segmens, arr)
 		return segmens
 	}
 	quantity := max / num
-	end := int64(0)
-	for i := int64(1); i <= num; i++ {
+	end := 0
+	for i := 1; i <= num; i++ {
 		qu := i * quantity
 		if i != num {
 			segmens = append(segmens, arr[i-1+end:qu])
